CustomEncryptor.py is the primary program file; the other 2 have functions it calls
These are the 3 program files that can be run with CustomEncryptor.

CustomEncryptor.py is the main program file. It's used to create ciphers for encryption, which can then be used to encrypt and decrypt data. These ciphers can be broken up and stored across multiple locations, thus increasing their security.

FetchEncryptedFiles.py is a program for easing the access to files in an encrypted zip folder on a flash drive. This is useful in CustomEncryptor because, by saving a portion of the cipher in an encrypted zip folder on a flash drive, one can essentially create a hardware key for any number of applications.

SortDictionary.py is completely optional. It's a simple program for sorting `dict` objects alphabetically and numerically. I've included it here because it is helpful for creating and checking ciphers. However, it is not required for CustomEncryptor to function.